requirements.txt
apache-airflow==2.9.3
flask==2.2.5
flask-restx==1.1.0
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
python-dotenv==1.0.1
mlflow==2.12.1
matplotlib==3.7.3